bug fixes following merge
